lava_dispatcher: drop hard guestfs dependency
The `guestfs` module is only required when emulating hardware. For
workers which only handle external devices this shouldn't be a hard
requirement.

With this commit the `guestfs` module is only imported when acutally
using the functions. Since `import` is smart it won't import modules
again after the initial run, not slowing down the system.

Signed-off-by: Paul Spooren <mail@aparcar.org>